WIP: Implement confounders menu above plot
